feat: SearchToolViewController 별도 폴더에 구현
#15
